requirements.txt
dvc
dvc[gdrive]
pandas
kaggle
numpy
pyyaml
openpyxl
sklearn
python-docx
python3-discogs-client
tqdm